+++ src/router/App.jsx
@@ -4,6 +4,7 @@ import AuthDetails from '../AuthDetails';
 import Home from '../pages/Home';
 import '../App.css';
 import {BrowserRouter as Router, Route, Switch} from 'react-router-dom'
+import Submission_Portal from '../pages/Submission_Portal';
 
 function App() {
   return (
@@ -14,10 +15,12 @@ function App() {
             <Route path='/home'>
               <Home/>
             </Route>
-
             <Route exact path='/'>
               <Login/>
               <AuthDetails/>
+            </Route>
+            <Route path='/submission_portal'>
+              <Submission_Portal/>
             </Route>
 
           </Switch>
